filter tests added, luma bug fixed

+++ src/model/filters/Darken_Luma.java
@@ -10,8 +10,8 @@ public class Darken_Luma implements Filter {
 
     // modifies every pixel in the layer
     for (int i = 0; i < layer.getTotalPixels(); i++) {
-      int luma = (int) (.2126 * layerToModify[i][0]) +
-          (int) (.7152 * layerToModify[i][1]) + (int) (.0722 * layerToModify[i][2]);
+      int luma = (int) (.2126 * layerToModify[i][0] +
+           (.7152 * layerToModify[i][1]) +  (.0722 * layerToModify[i][2]));
 
       for (int j = 0; j < 3; j++) {
         //adds max value to pixel without going over the cap
